[REFACTOR] task modal into its own component and [ADD] useImperativeHandle
